requirements.txt
numpy==1.24.3
scipy==1.11.1
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
flask==2.3.2
plotly==5.15.0
dash==2.11.1
dash-bootstrap-components==1.4.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0